extract useAuth hook to own module
